requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pypdf
pymongo
PyJWT
pymongo
bson
